Fixed a bug in font loading
When loading a font it could be that some chars aren't contained in the font, so the char width is 0 and the constructor of BufferedImage throws an exception. Now the char width gets checked before creating the image.
Also removed enable blending in the texture example states as it is already set when initializing the renderer.